+++ app/hooks/phone/index.ts
@@ -1,2 +1,3 @@
 export { usePhoneNumbers } from './usePhoneNumbers';
+export { useWorkspaceNumberSettingsMutations } from './useWorkspaceNumberSettingsMutations';
