+++ app/lib/platform-onboarding-helpers.server.ts
@@ -1,22 +1,29 @@
 import {
   getWorkspacePhoneNumbers,
 } from "@/lib/database/workspace.server";
+import { getWorkspaceRecentOutboundMessageCount } from "@/lib/database/workspace-twilio-portal-snapshot.server";
 import type { Tables } from "@/lib/db-types";
 import {
   applyOnboardingStepsWithWorkspaceNumbers,
   applyWorkspaceOnboardingChannelPolicy,
+  deriveWorkspaceMessagingReadiness,
   getWorkspaceMessagingOnboardingState,
 } from "@/lib/messaging-onboarding.server";
 import type { OnboardingActionData } from "@/lib/onboarding-actions.server";
 import {
+  getWorkspaceRcsBlockingIssues,
   hydrateWorkspaceRcsOnboardingState,
+  isRcsOnboardingEnabled,
 } from "@/lib/rcs-onboarding.server";
+import { buildA2pBlockingIssues } from "@/lib/twilio-a2p.server";
 import type {
   WorkspaceMessagingOnboardingState,
   WorkspaceMessagingReadiness,
   WorkspaceOperatingCountry,
 } from "@/lib/types";
 import { WORKSPACE_OPERATING_COUNTRY_VALUES } from "@/lib/types";
+import { getWorkspaceCredits } from "@/lib/workspace-members-db.server";
+import { createTenantDb } from "@/server/tenant-db";
 
 export type OnboardingHandlerResult =
   | {
@@ -50,6 +57,19 @@ export type WorkspaceOnboardingDetail = {
   rcs_blocking_issues: string[];
   phone_numbers: Tables<"workspace_number">[] | null;
   credits_balance: number;
+};
+
+export type WorkspaceOnboardingView = {
+  onboarding: WorkspaceMessagingOnboardingState;
+  phoneNumbers: Tables<"workspace_number">[] | null;
+  creditsBalance: number;
+  readiness: WorkspaceMessagingReadiness;
+  rcsBlockingIssues: string[];
+  a2pBlockingIssues: string[];
+  audienceCount: number;
+  campaignCount: number;
+  scriptCount: number;
+  recentOutboundCount: number;
 };
 
 function jsonInputToFormData(body: Record<string, unknown>): FormData {
@@ -101,42 +121,42 @@ export function resolveOperatingCountryFromForm(
     : fallback;
 }
 
-/** Prefer an in-workspace returnTo path; otherwise return a success payload. */
+/** Prefer an in-workspace returnTo path; otherwise return a success or error payload. */
 export function redirectToReturnToOrPayload(
   formData: FormData,
   workspaceId: string,
   savedKey: string,
   successMessage: string,
+  options?: { error?: string; status?: number },
 ): OnboardingHandlerResult {
   const returnTo = String(
     formData.get("returnTo") ?? formData.get("return_to") ?? "",
   );
   if (returnTo.startsWith(`/workspaces/${workspaceId}`)) {
+    if (options?.error) {
+      return {
+        kind: "redirect_path",
+        path: returnTo,
+        searchParams: { warning: options.error },
+      };
+    }
     return {
       kind: "redirect_path",
       path: returnTo,
       searchParams: { saved: savedKey },
     };
   }
+  if (options?.error) {
+    return {
+      kind: "payload",
+      data: { error: options.error },
+      status: options.status,
+    };
+  }
   return {
     kind: "payload",
     data: { success: successMessage },
   };
-}
-
-export function adaptRouteDataResult(result: unknown): OnboardingHandlerResult {
-  if (result && typeof result === "object" && "data" in result) {
-    const wrapped = result as {
-      data: OnboardingActionData;
-      init?: number | { status?: number } | null;
-    };
-    const status =
-      typeof wrapped.init === "number"
-        ? wrapped.init
-        : wrapped.init?.status ?? 200;
-    return { kind: "payload", data: wrapped.data, status };
-  }
-  return { kind: "payload", data: {}, status: 200 };
 }
 
 export async function hydrateWorkspaceOnboarding(workspaceId: string) {
@@ -151,4 +171,67 @@ export async function hydrateWorkspaceOnboarding(workspaceId: string) {
   );
 
   return { phoneNumbers: phoneNumbers ?? null, onboarding: hydratedOnboarding };
+}
+
+export async function loadWorkspaceOnboardingView(
+  workspaceId: string,
+): Promise<WorkspaceOnboardingView> {
+  const tdb = createTenantDb(workspaceId);
+
+  const [
+    { data: phoneNumbers },
+    onboarding,
+    creditsBalance,
+    recentOutboundCount,
+    audienceCount,
+    campaignCount,
+    scriptCount,
+  ] = await Promise.all([
+    getWorkspacePhoneNumbers({ workspaceId }),
+    getWorkspaceMessagingOnboardingState({ workspaceId }),
+    getWorkspaceCredits(workspaceId),
+    getWorkspaceRecentOutboundMessageCount({ workspaceId }),
+    tdb.audience.count(),
+    tdb.campaign.count(),
+    tdb.script.count(),
+  ]);
+
+  const hydratedOnboarding = applyOnboardingStepsWithWorkspaceNumbers(
+    hydrateWorkspaceRcsOnboardingState(applyWorkspaceOnboardingChannelPolicy(onboarding)),
+    phoneNumbers ?? [],
+    {
+      audienceCount,
+      scriptCount,
+      campaignCount,
+      creditsBalance: creditsBalance ?? 0,
+    },
+  );
+
+  const rcsBlockingIssues =
+    isRcsOnboardingEnabled() && hydratedOnboarding.selectedChannels.includes("rcs")
+      ? getWorkspaceRcsBlockingIssues(hydratedOnboarding)
+      : [];
+
+  const readiness = deriveWorkspaceMessagingReadiness({
+    onboarding: hydratedOnboarding,
+    workspaceNumbers: (phoneNumbers ?? []).map((number) => ({
+      type: number?.type ?? null,
+      phone_number: number?.phone_number ?? null,
+      capabilities: number?.capabilities ?? null,
+    })),
+    recentOutboundCount,
+  });
+
+  return {
+    onboarding: hydratedOnboarding,
+    phoneNumbers: phoneNumbers ?? null,
+    creditsBalance: creditsBalance ?? 0,
+    readiness,
+    rcsBlockingIssues,
+    a2pBlockingIssues: buildA2pBlockingIssues(hydratedOnboarding),
+    audienceCount,
+    campaignCount,
+    scriptCount,
+    recentOutboundCount,
+  };
 }
